requirements.txt
numpy>=1.26.5
pandas>=2.2.0
streamlit==1.48.1
streamlit-mic-recorder
pyarrow>=14.0.0
python-multipart>=0.0.9
fastapi>=0.110,<1
uvicorn[standard]>=0.23,<1
pydantic>=2.5,<3
python-dotenv>=1.0,<2
requests>=2.31,<3
openai==1.101.0
streamlit-drawable-canvas==0.9.3
sqlalchemy>=2.0,<3
aiosqlite>=0.19,<1
alembic>=1.13,<2
pdfplumber>=0.10.0,<1
python-docx>=1.1.0,<2
chardet>=5.0.0,<6
langchain_text_splitters==1.0.0
chromadb==1.3.6
pydantic-settings==2.12.0
asyncpg==0.31.0